movido de las views a su propia app lo relacionado a intervenciones.
